SpiderMap/requirements.txt
requests
dnspython
rich
